🗺️ Atlas: [Sub-modularizing the HeapFile Blob]
Applied the Facade pattern by extracting the physical page format structures (SlotEntry, VersionedSlotEntry, SlottedPage, VersionedSlottedPage, TupleId) and their associated serialization logic into a newly created page_format.rs internal sub-module. The mod.rs file now acts as an orchestrator, re-exporting these encapsulated internals to flatten the module boundaries, effectively isolating data structures from operational implementation.
